ccitt: style fixes
This commit applies some style fixes raised in the
https://go-review.googlesource.com/c/image/+/194417 "ccitt: set
bitReader default order to MSB" review.

Having nextBit return uint64, not uint32, removes some type conversions.

Removed a redundant "& 1" computation.

Updated a comment with s/LSB/MSB/, missed in the previous commit.

Restored the semantics of the "nBits" variable. Throughout the package,
there are a number of times where we use paired local variables or
paired struct fields, named "bits" and "nBits". The semantics is that
"bits"' type is uint32 or uint64 that actually holds a variable number
of bits, and "nBits" is that variable number. The previous commit,
writing MSB-first instead of LSB-first, changed the semantics (without
changing the variable name) so that one of those "nBits" variables now
held the index of the next bit, not the number of bits. This commit
restores the semantics: the shift is (7 - nBits), not just nBits.

Change-Id: I9afada9becceeb5642ce2c60eaf7bc0ede0cdd12
Reviewed-on: https://go-review.googlesource.com/c/image/+/199957
Reviewed-by: Horst Rutter <hhrutter@gmail.com>
Reviewed-by: Benny Siegert <bsiegert@gmail.com>
Run-TryBot: Benny Siegert <bsiegert@gmail.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>