image/ccitt: return all data from CCITT reader
The CCITT reader breaks out of the loop early when it's underlying
reader hits the io.EOF, even thought there was data still to be
processed.

This is because Read may return both bytes and an io.EOF error.

```
When Read encounters an error or end-of-file condition after
successfully reading n > 0 bytes, it returns the number of bytes read.
It may return the (non-nil) error from the same call or return the error
(and n == 0) from a subsequent call.
```
https://golang.org/pkg/io/#Reader

This test file, taken from libtiff (http://www.libtiff.org/images.html),
fails to be read properly with:

```
tiff: invalid format: not enough pixel data
```

Fixes golang/go#34809