bmp: ignore alpha in Decode when the alpha channel is not supported
When the alpha channel is not supported, the alpha component in the
original image should be ignored in the decoded image data (that is,
set to 0xFF in the decoded data).

We only support BITMAPFILEHEADER, BITMAPV4HEADER and BITMAPV5HEADER
with compression set to BIRBG or BITFIELDS as of now.
Among the combinations, BITMAPFILEHEADER does not support the alpha
channel while BITMAPV4HEADER and BITMAPV5HEADER does.

testdata/yellow_rose-small.png has been modified to pass the test.
No new test is added because testdata/yellow_rose-small.bmp has
covered this case.

References:

https://docs.microsoft.com/previous-versions/dd183376(v=vs.85)
https://en.wikipedia.org/wiki/BMP_file_format#Pixel_format